condensed scenes into 1 scene and added delegate dismissal of game scene
