+++ Lab3/Commotion/Base.lproj/Main.storyboard
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="15400" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="uAs-2Q-GOu">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="15400" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="5T7-yh-7UU">
     <device id="retina4_7" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
@@ -56,6 +56,37 @@
             </objects>
             <point key="canvasLocation" x="565" y="337"/>
         </scene>
+        <!--View Controller-->
+        <scene sceneID="SfV-er-wDU">
+            <objects>
+                <viewController id="5T7-yh-7UU" sceneMemberID="viewController">
+                    <layoutGuides>
+                        <viewControllerLayoutGuide type="top" id="tkT-Za-cdb"/>
+                        <viewControllerLayoutGuide type="bottom" id="AHw-dI-u7x"/>
+                    </layoutGuides>
+                    <view key="view" contentMode="scaleToFill" id="mfP-ft-J2H">
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="PRh-0k-R5s">
+                                <rect key="frame" x="146" y="318" width="82" height="30"/>
+                                <state key="normal" title="Go to Game"/>
+                                <connections>
+                                    <segue destination="uAs-2Q-GOu" kind="presentation" id="mU4-m8-fwE"/>
+                                </connections>
+                            </button>
+                        </subviews>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
+                        <constraints>
+                            <constraint firstItem="PRh-0k-R5s" firstAttribute="centerY" secondItem="mfP-ft-J2H" secondAttribute="centerY" id="DeB-Ak-dEl"/>
+                            <constraint firstItem="PRh-0k-R5s" firstAttribute="centerX" secondItem="mfP-ft-J2H" secondAttribute="centerX" id="m8o-IH-YDM"/>
+                        </constraints>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="irr-YJ-OhC" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="-388" y="978"/>
+        </scene>
         <!--Game View Controller-->
         <scene sceneID="rB5-LK-c32">
             <objects>
@@ -65,7 +96,7 @@
                         <viewControllerLayoutGuide type="bottom" id="9ra-ia-fff"/>
                     </layoutGuides>
                     <view key="view" contentMode="scaleToFill" id="9m8-Lp-cfY" customClass="SKView">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="647"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                     </view>
